requirements.txt
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.87.0
Pillow==11.2.0
pytesseract==0.3.11
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
stripe==12.2.0
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
